Make standard form's search fields sortable
In search page configuration you could choose which search fields to
enable, and it would result in a dropdown list in the search form. But
you could not choose the order in which fields appear in that list.
This patch fixes that by using the same technique as for the facet
fields and sort fields.
It also switch from jQueryUI's sortable to SortableJS, which is
distributed with Omeka S.
jQueryUI is still included because its "dialog" feature is still used