Change UI of search page config form for facet, sort and search fields
This change will allow to add more configuration elements to each field.

It also reduces visual noise when there are a lot of available fields
but not many enabled fields, and it looks more "Omeka-like"
as it takes (heavy) inspiration from the "Columns" form element.

Configuration for each field is now done in the sidebar instead of
jQuery UI's dialog, which means we no longer need jQuery UI