Add new plugin manager for facet value renderers
Facet value renderers change how a facet value is displayed to users.
They are useful when the indexed value is an internal id or code that
can be associated to a human-friendly text.
Omeka S modules can add their own facet value renderers.

This patch comes with 2 facet value renderers:
- 'fallback' which renders the facet value as is (current behaviour).
  It's the default renderer and is used whenever the configured value
  renderer cannot be found (when a module is disabled for instance)
- 'resource_title', which takes a resource ID as input and returns the
  corresponding resource title (displayTitle)