Always index resources in a separate process
Indexing resources in the same process as the one modifying resources
can cause a lot of different issues: performance, memory usage,
conflicting use of the Doctrine entity manager causing errors or data
duplication, ...

Doing it in a separate process ensure that we do not disturb the main
process.

There is a new background job (`Search\Job\Sync`) that is run
periodically when there are resources that need to be indexed.
Alternatively, there is also a new script (`bin/sync`) that can be
executed periodically (using cron for instance). It does the same thing
as the new background job.

Also, show the latest indexation date and time on admin resource details
page